remove redundant code, refactor
